refactor: introduce validation package
