+++ cognite/neat/rules/validation/_base.py
